[add_roles_management]: request and response dtos moved
