+++ src/main/java/com/carry/basar/model/dto/auth/AuthResponse.java
@@ -1,4 +1,4 @@
-package com.carry.basar.model;
+package com.carry.basar.model.dto.auth;
 
 public class AuthResponse {
     private final String jwt;
